+++ NEWS.md
@@ -2,6 +2,7 @@ Major news relating to breaking changes in ApproxManifoldProducts.jl
 
 ## v0.3 
 
+- Upgrade to ManifoldsBase.jl v0.11 with `AbstractManifold`.
 - Start consolidating internal manifold definitions to Manifolds.jl definitions instead.
 
 ## v0.2